backend/requirements.txt
websocket-client
python-dotenv
sounddevice
numpy
requests
fastapi
uvicorn
webrtcvad
chromadb
sentence-transformers
onnxruntime
sherpa-onnx
piper-tts
psutil
pytest
